refactor for separation of concerns, add observer pattern to loop

+++ src/main/groovy/com/groovygame/util/Time.groovy
@@ -4,7 +4,7 @@ import groovy.transform.Immutable
 
 @Immutable
 class Time {
-    static long getCurrentMilliseconds() {
+    static int getCurrentMilliseconds() {
         new Date().getTime()
     }
 }
